Rename format helper and update imports
